Major redesign. Put the pieces together.
- Add a queue to the timer. It feeds the periodic uip calls, and the readout of the BMX160 sensor.
- I2C.c can process asynchonous transfers. You can provide a callback function pointer when the transfer is finished.
Prerequisite are continuous calls to I2CPoll.
- BMX160 continuous reads are now timer driven, and fully asynchonous. It distinguishes between accel and gyro only and accel, gyro, and mag data readouts depending on the status. Status is read (also asynchonously) until at least new accel and gyro data are present.

Todo: Main loop body should run only when real data are present, not when internal processing of serial and I2C data awakens the processor.
Perform a sodt reset of BMX160 before (re)-configuring it. Currently errors must be ignored because the initial status of the sensor only happens after a physical power-up. When resetting the MCU, or re-programming it, or starting debugging the initialization reports errors.