I2C: Remove the driver task altogether, set the clock speed correctly
Eliminate the driver task altogether and move the I2C state machine back
into the interrupt handler.
It turned out that the task switch for each character to send was 1-2 ms
which immensly slowed down the net transfer rate.
Now the transfer is comforable within the 20ms frame including sending
the data via PPP through the USART.

Setting the I2C clock rate was missing altogether. Ups!
It is pure coincidence that the circuit ran at "only" 500Hz, and that
the BMX160 played along.